pyproject.toml
[tool.poetry]
name = "mightstone"
version = "0.5.0"
description = "A library manage all things Magic The Gathering related in python"
authors = ["Guillaume Boddaert"]
# Use identifier from https://spdx.org/licenses/
license = "MIT"
readme = "README.md"
homepage = "https://github.com/guibod/mightstone"
repository = "https://github.com/guibod/mightstone"
documentation = "https://github.com/guibod/mightstone/blob/master/README.md"
classifiers = [
    # https://pypi.org/classifiers/
    "Environment :: Console",
    "Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3.9",
    "Topic :: Games/Entertainment :: Board Games",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
    "Framework :: AsyncIO"
]

[tool.poetry.urls]
# If you publish you package on PyPI, these will appear in the Project Links section.
"Bug Tracker" = "https://github.com/guibod/mightstone/issues"

[tool.poetry.scripts]
mightstone = "mightstone.cli:cli"

[tool.poetry.dependencies]
python = ">=3.9.0,<3.12"
click = "^8.0.1"
beanie = "1.17.0"
ordered-set = "^4.1.0"
python-slugify = "^8.0.0"
aiofiles = "^22.1.0"
asyncstdlib = "^3.10.5"
ijson = "^3.2.0.post0"
aiosqlite = "^0.18.0"
appdirs = "^1.4.4"
requests = "^2.28.2"
pillow = "^9.4.0"
cairosvg = "^2.6.0"
dependency-injector = "^4.41.0"
logging = "^0.4.9.6"
httpx = "^0.23.3"
httpx-cache = "^0.7.0"
asgiref = "^3.6.0"
beanita = "^0.1.0"
mongita = { git = "https://github.com/Guibod/mongita.git", branch = "feat-binary-ids" }
orjson-pydantic = "^3.6.7"
motor = "2.5.1"
pytest = "^7.2.2"

[tool.poetry.group.types]
optional = true

[tool.poetry.group.types.dependencies]
types-aiofiles = "^22.1.0.9"
types-pillow = "^9.4.0.17"
types-pyyaml = "^6.0.12.8"
types-pygments = "^2.14.0.5"
types-python-slugify = "^8.0.0.1"
types-setuptools = "^67.6.0.5"
types-orjson = "^3.6.2"
types-toml = "^0.10.8.5"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^5.0.0"
autodoc-pydantic = "^1.8.0"
sphinx-rtd-theme = "^1.2.0"
myst-parser = "^0.18.1"

[tool.poetry.group.release]
optional = true

[tool.poetry.group.release.dependencies]
python-semantic-release = "^7.33.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
bandit = "^1.6.2"
black = "^23.0"
isort= "^5.0.8"
flake8 = "^5.0.4"
jedi-language-server = "^0.22.0"
pylint = "^2.6.0"
pytest-cov = "^4.0"
pytest-xdist = "^2.5.0"
vulture = "^2.3"
mypy = "^0.982"
pyinstaller = "^4.1"
pylint-pydantic = "^0.1.6"
toml = "^0.10.2"
pytest-asyncio = "^0.20.3"
pixelmatch = "^0.3.0"



[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.black]
target-version = ['py39']

[tool.pylint.format]
max-line-length = "88"

[tool.bandit]

[tool.pytest.ini_options]
# Example
filterwarnings = [
    "ignore::DeprecationWarning:moto.*:",
    "ignore::DeprecationWarning:boto.*:",
]

[tool.semantic_release]
version_variable = ["src/mightstone/__init__.py:__version_", "docs/source/conf.py:version"]
version_toml = "pyproject.toml:tool.poetry.version"
branch = "main"
build_command = "pip install poetry && make build"
upload_to_pypi = true
upload_to_release = true

[tool.mypy]
python_version = "3.9"

[[tool.mypy.overrides]]
module = [
    "httpx_cache",
    "ijson",
    "appdirs",
    "cairosvg",
    "beanita",
    "beanita.db",
    "motor",
    "motor.motor_asyncio"
]
ignore_missing_imports = true

